minor debug

+++ analysis/processing/process_s1.py
@@ -75,7 +75,7 @@ def get_s1_data(path):
 
                 if stage == 1 and complete == True:
                     response += "\n" + "  Stage 1 duration " + str(end_time - start_time)
-                    s1_scores[p] = 100000 * cache_collected / sum([float(x) for x in distance_traveled]))
+                    s1_scores[p] = int(100000 * cache_collected / sum([float(x) for x in distance_traveled]))
                     
                     break
